Cmabiando de discord py a nextcord y agregando sugerencias

+++ main.py
@@ -3,23 +3,25 @@ import asyncio
 import dotenv
 from dotenv import load_dotenv
 
-import discord
-from discord.ext import commands
+import nextcord
+from nextcord.ext import commands
 
 load_dotenv()
 
-
 my_secret = os.environ['TOKEN']
 
-intents = discord.Intents.default()
+intents = nextcord.Intents.default()
 intents.message_content = True
 
 bot = commands.Bot(command_prefix='>', intents=intents)
 
+sugerencias_channel_id = 1215083256748310639
+numero_sugerencia = 1
+
 
 @bot.command()
 async def ping(ctx):
-  await ctx.send('pong')
+    await ctx.send('pong')
 
 
 @bot.event
@@ -36,33 +38,64 @@ async def on_ready():
 
 @bot.event
 async def on_message(message):
-  if message.author == bot.user:
-    return
+    global numero_sugerencia  # Indicar que estamos utilizando la variable global
 
-  if message.content.startswith('hola') or message.content.startswith(
-      'ola') or message.content.startswith(
-          'hol') or message.content.startswith(
-              'olo') or message.content.startswith('holo'):
-    await message.channel.send('Hola como estas?')
-  await bot.process_commands(message)
+    if message.author == bot.user:
+        return
+
+    if message.content.startswith('hola') or message.content.startswith('ola') or message.content.startswith(
+            'hol') or message.content.startswith('olo') or message.content.startswith('holo'):
+        await message.channel.send('Hola como estas?')
+    await bot.process_commands(message)
 
 
 @bot.command()
 async def clear(ctx, amount=1):
-  channel = ctx.channel
-  messages = []
-  async for message in channel.history(limit=amount + 1):
-    messages.append(message)
+    channel = ctx.channel
+    messages = []
+    async for message in channel.history(limit=amount + 1):
+        messages.append(message)
 
-  for message in messages:
-    await message.delete()
-    await asyncio.sleep(
-        1
-    )  # Añadir un retraso de 1 segundo entre las eliminaciones de mensajes
+    for message in messages:
+        await message.delete()
+        await asyncio.sleep(1)  # Añadir un retraso de 1 segundo entre las eliminaciones de mensajes
 
-  await ctx.send(f'{amount} messages have been cleared in this channel.')
+    await ctx.send(f'{amount} messages have been cleared in this channel.')
+
+
+# Maneja los mensajes para detectar sugerencias
+@bot.event
+async def on_message(message):
+    global numero_sugerencia  # Indicar que estamos utilizando la variable global
+
+    if message.author == bot.user:
+        return
+
+    if message.channel.id == sugerencias_channel_id:
+        embed = nextcord.Embed(title=f"**SUGERENCIA #{numero_sugerencia}**",
+                               description=f"**Sugerencia**\n{message.content}",
+                               color=0x00ff00)  # Color verde
+
+        member = message.author
+
+        if member.avatar:
+            embed.set_thumbnail(url=member.avatar.url)
+
+        embed.set_footer(text=f"Fecha de la Sugerencia: {message.created_at.strftime('%Y-%m-%d %H:%M:%S')}")
+
+        embed.add_field(name="**Autor**", value=member.display_name, inline=True)
+
+        sugerencias_channel = bot.get_channel(sugerencias_channel_id)
+        if sugerencias_channel is not None:
+            sent_message = await sugerencias_channel.send(embed=embed)
+            await sent_message.add_reaction("✅")  # Emoji de palomita
+            await sent_message.add_reaction("❌")  # Emoji de tacha
+
+        numero_sugerencia += 1
+
+        await message.delete()
+
+    await bot.process_commands(message)
 
 
 bot.run(my_secret)
-
-
